requirements.txt
asgiref==3.8.1
chardet==5.2.0
Django==5.2.1
django-quill-editor==0.1.42
djangorestframework==3.16.0
et_xmlfile==2.0.0
openpyxl==3.1.5
pillow==11.2.1
reportlab==4.4.0
sqlparse==0.5.3
tzdata==2025.2
